requirements.txt
# Core AI/LLM Libraries
openai>=1.0.0
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-google-genai>=0.0.5
langgraph>=0.0.20

# Vector Databases & Embeddings
faiss-cpu>=1.7.4
langchain-google-genai>=0.0.5
sentence-transformers>=2.2.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Research & Data Processing
arxiv>=2.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pypdf>=3.17.0

# Database & Memory
chromadb>=0.4.0
pinecone-client>=3.0.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0
